+++ skywalking/agent/protocol/kafka.py
@@ -53,7 +53,7 @@ class KafkaProtocol(Protocol):
 
             while True:
                 try:
-                    timeout = config.QUEUE_TIMEOUT  # type: int
+                    timeout = config.queue_timeout  # type: int
                     if not start:  # make sure first time through queue is always checked
                         start = time()
                     else:
@@ -117,7 +117,7 @@ class KafkaProtocol(Protocol):
 
             while True:
                 try:
-                    timeout = config.QUEUE_TIMEOUT  # type: int
+                    timeout = config.queue_timeout  # type: int
                     if not start:  # make sure first time through queue is always checked
                         start = time()
                     else:
@@ -144,7 +144,7 @@ class KafkaProtocol(Protocol):
 
             while True:
                 try:
-                    timeout = config.QUEUE_TIMEOUT  # type: int
+                    timeout = config.queue_timeout  # type: int
                     if not start:  # make sure first time through queue is always checked
                         start = time()
                     else: