+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 ## Change Logs
 
 ### 1.0.0
+
+- **Important Note and Breaking Changes:**
+  - Python 3.6 is no longer supported and may not function properly.
+  - A number of configuration options (mostly environment variables) are renamed to follow the convention of
+  'SW_AGENT' prefixes, please check with the latest official documentation before upgrading. (#273)
+  
+  https://skywalking.apache.org/docs/skywalking-python/v1.0.0/en/setup/configuration/
+  - All agent core capabilities are now covered by test cases and enabled by default (Trace, Log, PVM runtime metrics, Profiler)
+
+
 - Feature:
-  - Drop support for Python 3.6
-  - Add support for Python 3.11 (Pending)
-  - Add MeterReportService (gRPC, Kafka reporter) (default:disabled) (#231, #236, #241, #243)
-  - Add reporter for PVM runtime metrics (default:disabled) (#238, #247)
+  - Add MeterReportService (gRPC, Kafka reporter) (default:enabled) (#231, #236, #241, #243)
+  - Add reporter for PVM runtime metrics (default:enabled) (#238, #247)
   - Add Greenlet profiler (#246)
   - Add test and support for Python Slim base images (#249)
   - Add support for the tags of Virtual Cache for Redis (#263)
@@ -29,6 +37,7 @@
   - Enforce tag class type conversion (#262)
   - Fix sw_logging (log reporter) potentially throw exception leading to traceback confusion (#267)
   - Avoid reporting meaningless tracecontext with logs when there's no active span, UI will now show empty traceID (#272)
+  - Fix exception handler in profile_context (#273)
   
 - Docs:
   - New documentation on how to test locally (#222)
@@ -36,6 +45,7 @@
   - New documentation on the newly added greenlet profiler and the original threading profiler (#250)
   - Overhaul documentation on development setup and testing (#249)
   - Add tables to state currently supported features of Python agent. (#271)
+  - New configuration documentation generator (#273) 
 
 - Others:
   - Pin CI SkyWalking License Eye (#221)